server/requirements.txt
flask==2.1.0
flask-cors==3.0.10
python-dotenv==0.19.2
pymongo==4.1.1
python-jose==3.3.0
requests==2.27.1
six==1.16.0